requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
asgiref==3.6.0
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
cmake==3.26.1
coloredlogs==15.0.1
commentjson==0.9.0
contourpy==1.0.7
cycler==0.11.0
numpy==1.23.5
Django==4.0.10
django-cors-headers==3.14.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
fastapi==0.87.0
filelock==3.10.7
filetype==1.2.0
fire==0.5.0
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
ftfy==6.1.1
h11==0.14.0
humanfriendly==10.0
humanize==4.6.0
idna==3.4
ImageHash==4.3.1
imageio==2.27.0
importlib-resources==5.12.0
Jinja2==3.1.2
kiwisolver==1.4.4
lark-parser==0.7.8
lit==16.0.0
llvmlite==0.39.1
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
multidict==6.0.4
networkx==3.0
numba==0.56.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnxruntime==1.13.1
opencv-python-headless==4.6.0.66
packaging==23.0
Pillow==9.3.0
plotly==5.14.0
pooch==1.6.0
protobuf==4.22.1
pybind11==2.10.4
pydantic==1.10.7
PyMatting==1.1.8
pymongo==4.3.3
pyparsing==3.0.9
pyquaternion==0.9.9
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.3.23
rembg==2.0.30
removebg==0.4
requests==2.28.2
scikit-image==0.19.3
scipy==1.9.3
six==1.16.0
sniffio==1.3.0
sqlparse==0.2.4
starlette==0.21.0
sympy==1.11.1
tenacity==8.2.2
termcolor==2.2.0
tifffile==2023.3.21
torch==2.0.0
torchvision==0.15.1
tqdm==4.64.1
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.20.0
watchdog==2.1.9
wcwidth==0.2.6
yarl==1.8.2
pytest==7.2.2
